dom/messagechannel/tests/test_messageChannel_selfTransferring.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=677638
-->
<head>
  <meta charset="utf-8">
  <title>MessagePort/Channel no self tranferring</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=677638">Mozilla Bug 677638</a>
<div id="content"></div>
<pre id="test">
</pre>
  <script type="application/javascript">

  var a = new MessageChannel();

  var status = false;
  try {
    a.port1.postMessage('foobar', [a.port1]);
  } catch(e) {
    status =true;
  }

  ok(status, "Transfering the same port should throw");

  </script>
</body>
</html>
